# swagger.yaml
swagger: "2.0"
info:
  version: "1.0"
  title: "Local Shop Advisor"
  description: Documentazione delle API del progetto del corso di Ingegneria del Software
host: localhost:8080
basePath: /api/v1
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /products:
    get:
      summary: Lista di tutti i prodotti
      responses:
        200:
          description: Risposta elaborata correttamente
          schema:
            type: array
            items:
              $ref: "#/definitions/Product"
        404:
          description: Error not found
    post:
      consumes:
        - multipart/form-data
        - application/json
      summary: Creazione di un nuovo prodotto
      parameters:
        - in: body
          name: product
          required: true
          description: I parametri del prodotto da creare + immagini da caricare
          schema:
            $ref: "#/definitions/Product"
      responses:
        201:
          description: OK
          schema:
            $ref: "#/definitions/Product"
          examples:
            application/json:
              name: Felpa
              description: Una bella felpa
              price: 10
              category: Abbigliamento
              tags:
                - Abbigliamento
                - Felpe
              images:
                - images//image.jpg
                - images//image2.jpg
        404:
          description: Error not found
  /products/{productId}:
    get:
      summary: Get di un singolo prodotto
      parameters:
        - in: path
          name: productId
          required: true
          type: string
          description: L'id del prodotto da cercare
      responses:
        200:
          description: Risposta elaborata correttamente
          schema:
            $ref: "#/definitions/Product"
        404:
          description: Error not found
    delete:
      summary: Eliminazione di un singolo prodotto
      parameters:
        - in: path
          name: productId
          required: true
          type: string
          description: L'id del prodotto da eliminare
      responses:
        204:
          description: Prodotto rimosso correttamente
        400:
          description: Cannot delete images
        404:
          description: Error not found
    put:
      summary: Modifica di un prodotto
      parameters:
        - in: path
          name: productId
          required: true
          type: string
          description: L'id del prodotto da modificare
        - in: body
          name: product
          required: true
          description: I nuovi valori del prodotto da modificare
          schema:
            $ref: "#/definitions/Product"
      responses:
        200:
          description: Prodotto modificato correttamente
          schema:
            $ref: "#/definitions/Product"
        404:
          description: Error not found
  /products/find/{name}:
    get:
      summary: Ricerca di un prodotto per nome
      parameters:
        - in: path
          name: name
          required: true
          type: string
          description: Il nome del prodotto da cercare
      responses:
        200:
          description: Risposta elaborata correttamente
          schema:
            $ref: "#/definitions/Product"
        404:
          description: Error not found
  /images/{id}:
    get:
      summary: Static route per le immagini
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: Id dell'immagine da ottenere
      produces:
        - image/png
        - image/jpg
        - image/jpeg
      responses:
        200:
          description: Immagine
          schema:
            type: file
        404:
          description: Error not found
definitions:
  Product:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      price:
        type: number
        minimum: 0
        maximum: 5000
      category:
        type: string
      tags:
        type: array
        items:
          type: string
      images:
        type: array
        items:
          type: string
    required:
      - name
      - description
      - price
      - category
    example:
      name: Felpa
      description: Una bella felpa
      price: 10
      category: Abbigliamento
      tags:
        - Abbigliamento
        - Felpe
      images:
        - images//image.jpg
        - images//image2.jpg